treewide: replace bare int / long / long long with fixed-width types
Project rule: every integer declaration in the tree uses fixed-
width types from <stdint.h> / <inttypes.h>. No bare int, long,
short, unsigned int, long long, etc. The rule applies uniformly
to library code, test files, and the C entry point.

Changes:

- tests/unity_minimal.h: TEST_ASSERT_EQ expands to int64_t
instead of long long; uses PRId64 for the printf format so
x86_64 Linux (where int64_t = long int, not long long) does
not trip -Werror=format=.

- tests/test_*.c: every `int main(void)` becomes
`int32_t main(void)`; every `static int s_xxx` counter
becomes `static int32_t`; every `int ctx_val =...`
marker becomes `int32_t`; every `(int)` cast in
TEST_ASSERT_EQ becomes `(int32_t)`; every `(long long)`
cast becomes `(int64_t)`; every `*(int*)ctx` dereference
becomes `*(int32_t*)ctx`.

- src/main.c: `int main` -> `int32_t main`. Surrounded with
a GCC pragma push/pop that silences -Wmain-return-type,
which is pattern-matched against the literal spelling
``int`` regardless of the underlying typedef. int32_t is
typedef'd to int on every supported platform so the
effective return type is unchanged.

- src/boot/vector_table.c: the `extern int main` forward
declaration follows the same pattern.

- tests/mocks/ra_sim_mmap.c: the fprintf diagnostic cast goes
from (unsigned long) to (uint64_t). On x86_64 Linux these
types are identical so the existing "%lx" format specifier
still matches.

- tests/test_ra_error_handler.c: static int32_t s_trap_hit.

Gates: 48/48 ctests pass, coverage 97.2% lines / 91.0% branches,
clang-tidy clean, cppcheck clean, doxygen 0 warnings,
cross-build 5996/56/5120 ELF unchanged.

+++ tests/test_ra_dma.c
@@ -17,14 +17,14 @@
 #include "ra_sim_mmap.h"
 #include "unity_minimal.h"
 
-static int s_complete_count    = 0;
-static int s_complete_last_ctx = 0;
+static int32_t s_complete_count    = 0;
+static int32_t s_complete_last_ctx = 0;
 
 static void reset_state(void)
 {
   ra_sim_mmap_reset();
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_mstp_init());
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_init());
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_mstp_init());
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_init());
   s_complete_count    = 0;
   s_complete_last_ctx = 0;
 }
@@ -33,7 +33,7 @@ static void stub_complete(void* ctx)
 {
   ++s_complete_count;
   if (ctx != nullptr) {
-    s_complete_last_ctx = *(int*)ctx;
+    s_complete_last_ctx = *(int32_t*)ctx;
   }
 }
 
@@ -44,7 +44,7 @@ static void test_init_marks_channels_free(void)
 
   for (uint8_t ch = 0U; ch < (uint8_t)k_ra_dma_channel_count; ++ch) {
     bool busy = true;
-    TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_channel_is_busy(ch, &busy));
+    TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_channel_is_busy(ch, &busy));
     TEST_ASSERT(!busy);
   }
   TEST_END("ra_dma_init: all channels free");
@@ -72,11 +72,11 @@ static void test_request_allocates_channel_zero(void)
   };
 
   uint8_t ch = 0xFFU;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_request(&req, &ch));
-  TEST_ASSERT_EQ((int)0, (int)ch);
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)0, (int32_t)ch);
 
   bool busy = false;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_channel_is_busy(ch, &busy));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_channel_is_busy(ch, &busy));
   TEST_ASSERT(busy);
   TEST_END("ra_dma_request: first allocation gets channel 0");
 }
@@ -95,7 +95,7 @@ static void test_request_rejects_bad_engine(void)
     .width    = k_ra_dmac_width_byte,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_err_not_supported, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_not_supported, (int32_t)ra_dma_request(&req, &ch));
   TEST_END("ra_dma_request: unsupported engine rejected");
 }
 
@@ -113,7 +113,7 @@ static void test_request_rejects_zero_count(void)
     .width    = k_ra_dmac_width_byte,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_err_invalid_arg, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_invalid_arg, (int32_t)ra_dma_request(&req, &ch));
   TEST_END("ra_dma_request: zero count rejected");
 }
 
@@ -123,7 +123,7 @@ static void test_request_null_ptrs(void)
   reset_state();
 
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_err_null_ptr, (int)ra_dma_request(nullptr, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_null_ptr, (int32_t)ra_dma_request(nullptr, &ch));
 
   uint8_t                buf[4] = {};
   const ra_dma_request_t req    = {
@@ -133,7 +133,7 @@ static void test_request_null_ptrs(void)
     .count    = 1U,
     .width    = k_ra_dmac_width_byte,
   };
-  TEST_ASSERT_EQ((int)k_ra_err_null_ptr, (int)ra_dma_request(&req, nullptr));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_null_ptr, (int32_t)ra_dma_request(&req, nullptr));
   TEST_END("ra_dma_request: null pointers rejected");
 }
 
@@ -151,16 +151,16 @@ static void test_release_round_trip(void)
     .width    = k_ra_dmac_width_byte,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_request(&req, &ch));
 
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_release(ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_release(ch));
 
   bool busy = true;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_channel_is_busy(ch, &busy));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_channel_is_busy(ch, &busy));
   TEST_ASSERT(!busy);
 
   /* Second release is an error. */
-  TEST_ASSERT_EQ((int)k_ra_err_invalid_state, (int)ra_dma_release(ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_invalid_state, (int32_t)ra_dma_release(ch));
   TEST_END("ra_dma_release: channel returns to free pool");
 }
 
@@ -168,7 +168,7 @@ static void test_release_bad_channel(void)
 {
   TEST_BEGIN("ra_dma_release: out-of-range rejected");
   reset_state();
-  TEST_ASSERT_EQ((int)k_ra_err_invalid_arg, (int)ra_dma_release(99U));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_invalid_arg, (int32_t)ra_dma_release(99U));
   TEST_END("ra_dma_release: out-of-range rejected");
 }
 
@@ -188,10 +188,10 @@ static void test_channel_exhaustion(void)
 
   for (uint8_t i = 0U; i < (uint8_t)k_ra_dma_channel_count; ++i) {
     uint8_t ch = 0U;
-    TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_request(&req, &ch));
+    TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_request(&req, &ch));
   }
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_err_no_mem, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_no_mem, (int32_t)ra_dma_request(&req, &ch));
 
   TEST_END("ra_dma_request: no-mem when all channels busy");
 }
@@ -213,12 +213,12 @@ static void test_sim_dma_memcpy_byte(void)
     .dst_inc  = true,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_request(&req, &ch));
 
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_sim_dma_memcpy(ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_sim_dma_memcpy(ch));
 
-  TEST_ASSERT_EQ((int)0x10, (int)dst[0]);
-  TEST_ASSERT_EQ((int)0x60, (int)dst[5]);
+  TEST_ASSERT_EQ((int32_t)0x10, (int32_t)dst[0]);
+  TEST_ASSERT_EQ((int32_t)0x60, (int32_t)dst[5]);
   TEST_END("ra_sim_dma_memcpy: byte transfer");
 }
 
@@ -239,12 +239,12 @@ static void test_sim_dma_memcpy_word(void)
     .dst_inc  = true,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_request(&req, &ch));
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_sim_dma_memcpy(ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_sim_dma_memcpy(ch));
 
-  TEST_ASSERT_EQ((long long)0xDEADBEEFU, (long long)dst[0]);
-  TEST_ASSERT_EQ((long long)0xCAFEBABEU, (long long)dst[1]);
-  TEST_ASSERT_EQ((long long)0x12345678U, (long long)dst[2]);
+  TEST_ASSERT_EQ((int64_t)0xDEADBEEFU, (int64_t)dst[0]);
+  TEST_ASSERT_EQ((int64_t)0xCAFEBABEU, (int64_t)dst[1]);
+  TEST_ASSERT_EQ((int64_t)0x12345678U, (int64_t)dst[2]);
   TEST_END("ra_sim_dma_memcpy: word transfer");
 }
 
@@ -253,7 +253,7 @@ static void test_sim_dma_complete_fires_callback(void)
   TEST_BEGIN("ra_sim_dma_complete: callback invoked with ctx");
   reset_state();
 
-  int                    ctx_val = 0xABCD;
+  int32_t                ctx_val = 0xABCD;
   uint8_t                buf[4]  = {};
   const ra_dma_request_t req     = {
     .engine      = k_ra_dma_engine_dmac,
@@ -265,11 +265,11 @@ static void test_sim_dma_complete_fires_callback(void)
     .ctx         = &ctx_val,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_request(&req, &ch));
 
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_sim_dma_complete(ch));
-  TEST_ASSERT_EQ((int)1, (int)s_complete_count);
-  TEST_ASSERT_EQ((int)0xABCD, (int)s_complete_last_ctx);
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_sim_dma_complete(ch));
+  TEST_ASSERT_EQ((int32_t)1, (int32_t)s_complete_count);
+  TEST_ASSERT_EQ((int32_t)0xABCD, (int32_t)s_complete_last_ctx);
   TEST_END("ra_sim_dma_complete: callback invoked with ctx");
 }
 
@@ -277,10 +277,10 @@ static void test_dma_request_without_init_fails(void)
 {
   TEST_BEGIN("ra_dma_request: not-initialised rejected");
   ra_sim_mmap_reset();
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_mstp_init());
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_mstp_init());
   /* Deinit explicitly so s_initialized is false. */
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_init());
-  TEST_ASSERT_EQ((int)k_ra_ok, (int)ra_dma_deinit());
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_init());
+  TEST_ASSERT_EQ((int32_t)k_ra_ok, (int32_t)ra_dma_deinit());
 
   uint8_t                buf[4] = {};
   const ra_dma_request_t req    = {
@@ -291,7 +291,7 @@ static void test_dma_request_without_init_fails(void)
     .width    = k_ra_dmac_width_byte,
   };
   uint8_t ch = 0U;
-  TEST_ASSERT_EQ((int)k_ra_err_not_initialized, (int)ra_dma_request(&req, &ch));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_not_initialized, (int32_t)ra_dma_request(&req, &ch));
   TEST_END("ra_dma_request: not-initialised rejected");
 }
 
@@ -300,9 +300,9 @@ static void test_channel_is_busy_bad_inputs(void)
   TEST_BEGIN("ra_dma_channel_is_busy: bad inputs");
   reset_state();
 
-  TEST_ASSERT_EQ((int)k_ra_err_null_ptr, (int)ra_dma_channel_is_busy(0U, nullptr));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_null_ptr, (int32_t)ra_dma_channel_is_busy(0U, nullptr));
   bool busy = false;
-  TEST_ASSERT_EQ((int)k_ra_err_invalid_arg, (int)ra_dma_channel_is_busy(99U, &busy));
+  TEST_ASSERT_EQ((int32_t)k_ra_err_invalid_arg, (int32_t)ra_dma_channel_is_busy(99U, &busy));
   TEST_END("ra_dma_channel_is_busy: bad inputs");
 }
 
@@ -315,7 +315,7 @@ static void test_dispatch_out_of_range(void)
   TEST_END("ra_dma_dispatch_complete: out-of-range is no-op");
 }
 
-int main(void)
+int32_t main(void)
 {
   test_init_marks_channels_free();
   test_request_allocates_channel_zero();
